refactor: desktop-capturer webrtcUtil
